Renamed set_retry_time to retry
I did this change because it's easier to understand although SQS doesn't
require to manually retry messages. That will also make tests easier
because the `Dummy` connector was only collecting the retried jobs
which had a different `retry_time`. By doing this change, the retry
implementation relies on the Connector child classes